+++ src/typings/selectfield.d.ts
@@ -1,0 +1,5 @@
+declare interface SelectFieldItem {
+  readonly value: string | number
+  readonly label: string
+  readonly ariaLabel?: string
+}