+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[NEW]** Add call and send icons
 - **[UPDATE]** update `Itinerary` colors
+- **[NEW]** Add SelectField component
 - [...]
 
 # v0.22.1 (22/02/2019)